Move tests, correct import, mark slow
This is a general and relatively simple cleanup of testing code.
Basically some imports had to be changed, some code has been moved
around and I've marked slow tests as slow.
